Fix sender key upload 422: widen chain_id from i32 to i64 for u32 WASM values

+++ crates/chatalot-common/src/api_types.rs
@@ -803,7 +803,7 @@ pub struct TransferCommunityOwnershipRequest {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct UploadSenderKeyRequest {
-    pub chain_id: i32,
+    pub chain_id: i64,
     pub distribution: serde_json::Value,
 }
 
@@ -812,7 +812,7 @@ pub struct SenderKeyDistributionResponse {
     pub id: Uuid,
     pub channel_id: Uuid,
     pub user_id: Uuid,
-    pub chain_id: i32,
+    pub chain_id: i64,
     pub distribution: serde_json::Value,
     pub created_at: String,
 }
